fix: users can include a comma delimited line of arguments and it will be evaluated
